Move `getWebAppLaunchUrl` and `isWebApp` to `app.ts`, improve API
