+++ web/packages/teleterm/src/ui/services/clusters/clustersService.ts
@@ -640,34 +640,3 @@ export function makeApp(source: tsh.App): App {
 
   return { ...source, addrWithProtocol };
 }
-
-/** Returns a URL that can be used to open the app in a browser. */
-export function getWebAppLaunchUrl(
-  app: tsh.App,
-  clustersService: ClustersService
-): string {
-  const { fqdn, publicAddr } = app;
-
-  const rootCluster = clustersService.findCluster(
-    routing.ensureRootClusterUri(app.uri)
-  );
-  const cluster = clustersService.findClusterByResource(app.uri);
-
-  const canCreateUrl =
-    rootCluster.proxyHost && fqdn && cluster?.name && publicAddr;
-
-  if (!canCreateUrl) {
-    return '';
-  }
-  return `https://${rootCluster.proxyHost}/web/launch/${fqdn}/${cluster.name}/${publicAddr}`;
-}
-
-export function isWebApp(app: tsh.App): boolean {
-  if (app.samlApp || app.awsConsole) {
-    return false;
-  }
-  return (
-    app.endpointUri.startsWith('http://') ||
-    app.endpointUri.startsWith('https://')
-  );
-}
